requirements.txt
opencv-python
pillow
torch
tritonclient[all]
torchvision